Fixed Bug, See Desc, Updated Makefile and README
Fixed Bug causing .wav and .au files to be unplayable. The bug was the
decoded AVFrame returned from FFmpeg_Decoder::decode_frame() had its
channel_layout set to 0 causing the resampler to not work and give
errors. Added a couple lines to the end of
FFmpeg_Decoder::decode_frame() where if AVFrame->channel_layout is 0, it
will fix it and set it accordingly.